+++ coders/mpc.c
@@ -1207,7 +1207,7 @@ static MagickBooleanType WriteMPCImage(const ImageInfo *image_info,Image *image,
     (void) FormatLocaleString(buffer,MagickPathExtent,"number-channels=%.20g "
       "number-meta-channels=%.20g channel-mask=0x%16lx\n",
       (double) image->number_channels,(double) image->number_meta_channels,
-      (unsigned long) image->channel_mask);
+      (ssize_t) image->channel_mask);
     (void) WriteBlobString(image,buffer);
     (void) FormatLocaleString(buffer,MagickPathExtent,
       "columns=%.20g rows=%.20g depth=%.20g\n",(double) image->columns,
